[최서윤-240725] feat :: feat :: 리뷰 RUD, 다이어리 C 기능 구현

+++ src/main/java/com/multi/toonGather/social/model/dto/WriterDTO.java
@@ -3,7 +3,10 @@ package com.multi.toonGather.social.model.dto;
 import lombok.Data;
 
 /**
- * Please explain the class!!
+ * 리뷰 작성자 정보를 담는 데이터 전송 객체 (DTO)
+ *
+ * 이 클래스는 리뷰 작성자의 기본 정보(사용자 번호, ID, 닉네임)와
+ * 프로필 이미지 경로를 포함합니다.
  *
  * @author : seoyun
  * @fileName : WriterDTO
@@ -11,7 +14,7 @@ import lombok.Data;
  */
 @Data
 public class WriterDTO {
-    private Long userNo;
+    private int userNo;
     private String userId;
     private String nickname;
     private String profileImagePath;
